Create MS3 Challenge Code
the Purpose of this repo is to design an application for a customer.
Instruction
You will need a SQLite-JDBC jar file to run the program: This is where to download the version I am using 
https://bitbucket.org/xerial/sqlite-jdbc/downloads/ The next step is running the main function.
I personally used intellij but eclipse can be use too. I used DB Browser for sqlite to view the database. 
It can be download at https://sqlitebrowser.org/ The DDL commands I used were
Database Name: Ms3Customer
Table Name: Records

Approach
the file that was given to me was a large file. I took an array that can parse csv files.I notice each column were separated by commas so I used the comma to split the string. I created a function to check if all the column items were there. Then I created a different class to create the table and create an add function to insert into the table. Overall, I parsed the csv file to make sure I got the right output then create the database and insert the correct information.